smaller extract_subgraphs file
